+++ Demo/src/main/java/com/paypal/android/ui/approveorder/ApproveOrderViewModel.kt
@@ -207,4 +207,10 @@ class ApproveOrderViewModel @Inject constructor(
             )
         }
     }
+
+    override fun onCleared() {
+        super.onCleared()
+
+        cardClient.removeObservers()
+    }
 }