+++ Demo/src/main/java/com/paypal/android/ui/paypalweb/PayPalWebViewModel.kt
@@ -154,4 +154,10 @@ class PayPalWebViewModel @Inject constructor(
             }
         }
     }
+
+    override fun onCleared() {
+        super.onCleared()
+
+        paypalClient.removeObservers()
+    }
 }