+++ Demo/src/main/java/com/paypal/android/ui/vaultcard/VaultCardViewModel.kt
@@ -204,4 +204,9 @@ class VaultCardViewModel @Inject constructor(
         }
         cardClient.presentAuthChallenge(activity, authChallenge)
     }
+
+    override fun onCleared() {
+        super.onCleared()
+        cardClient.removeObservers()
+    }
 }